Major clean-up of the code for the sake of more comfortable development
Basically refactored all the backend code into their separate folders,
left only index.js to serve as a simple start-up script to the root.

All modules were transformed into either classes or objects my reasoning
being that the prototype pattern was unnecessarily ugly and convoluted.

Classes very used when functions had a state and objects for staless
functions. I tried to keep the code the way it was even though some
methods should have been cut and separated into multiple pieces.

The project still needs some heavy refactoring and documentation for it
to be a more pleasant experience to use. Spider-web code is cool when you're
rapidly prototyping but for anyone wanting to use this for something serious
they'd have to do the same thing as I did to introduce testing etc.